[misc] reorganize main react components
