+++ src/pages/index.astro
@@ -1,10 +1,8 @@
 ---
 import { App } from '../App.tsx'
-import { SupabaseProvider } from '../contexts/supabase.tsx'
-import Layout from '../layouts/Layout.astro'
+import Layout from '../layouts/main.astro'
 ---
+
 <Layout title="Home Page">
-  <SupabaseProvider client:load>
-    <App client:load/>
-  </SupabaseProvider>
+  <App/>
 </Layout>
